fix: ask 권한연결,유저정보 불러오기 수정,이미지연결"

+++ src/main/java/yiu/aisl/devTogether/service/AskService.java
@@ -13,6 +13,7 @@ import yiu.aisl.devTogether.dto.AskResponseDto;
 import yiu.aisl.devTogether.exception.CustomException;
 import yiu.aisl.devTogether.exception.ErrorCode;
 import yiu.aisl.devTogether.repository.AskRepository;
+import yiu.aisl.devTogether.repository.UserRepository;
 
 
 import java.util.ArrayList;
@@ -26,6 +27,7 @@ import java.util.Optional;
 @Transactional
 public class AskService {
     private final AskRepository askRepository;
+    private final UserRepository userRepository;
 
 
     //ask 조회
@@ -41,15 +43,18 @@ public class AskService {
 
 
     //ask 사용자 등록
-    public Boolean create(User user,AskRequestDto.CreateDTO request) {
+    public Boolean create(String email, AskRequestDto.CreateDTO request) {
+        User user = findByEmail(email);
+        System.out.println(user);
 
-        if (request.getTitle() == null || request.getContents() == null || request.getAskCategory() == null) {
+
+        if (request.getTitle() == null || request.getContents() == null || request.getAskCategory() == null
+         )
+         {
             throw new CustomException(ErrorCode.INSUFFICIENT_DATA);
         }
 
         try {
-
-
 
             Ask ask = Ask.builder()
                     .user(user)
@@ -67,12 +72,19 @@ public class AskService {
         return true;
     }
 
+
+
     //ask 관리자 답변
-    public Boolean answer(Long askId, AskRequestDto.AnswerDTO request) {
+    public Boolean answer(String email, Long askId, AskRequestDto.AnswerDTO request) {
 
         RoleCategory role = RoleCategory.fromInt(request.getRole());
         //404 - id없음
         Ask ask = findByAskId(request.getAskId());
+        // 403 - 권한 없음
+        User user = findByEmail(email);
+        if(user.getRole() !=RoleCategory.MANAGER ){
+            throw  new CustomException(ErrorCode.NO_AUTH);
+        }
 
         if (role == RoleCategory.MANAGER) {
             ask.setAnswer(request.getAnswer());
@@ -92,9 +104,13 @@ public class AskService {
 
 
     //ask 삭제
-    public Boolean delete(AskRequestDto.DeleteDTO request) {
+    public Boolean delete(String email, AskRequestDto.DeleteDTO request) {
         //404 - id없음
         Ask ask = findByAskId(request.getAskId());
+        //403 - 권한 없음  >> 자기가 쓴 글이 아닌경우
+        if (!ask.getUser().getEmail().equals(email)) {
+            throw new CustomException(ErrorCode.NO_AUTH);
+        }
         try{
             askRepository.deleteById(request.getAskId());
             return true;
@@ -108,8 +124,9 @@ public class AskService {
 
 
     //ask 수정
-    public Boolean update(AskRequestDto.UpdateDTO request) {
+    public Boolean update( String email, AskRequestDto.UpdateDTO request) {
         AskCategory askCategory = AskCategory.fromInt(request.getAskCategory());
+        User user = findByEmail(email);
         //404 - id없음
         Ask ask = findByAskId(request.getAskId());
         //400 - 데이터 미입력
@@ -117,11 +134,19 @@ public class AskService {
             throw new CustomException(ErrorCode.INSUFFICIENT_DATA);
 
         }
-        //404 - 권한 없음  >> 관리자가 답변을 달아 완료가 된 상태인데 수정하려는 경우
+        //404 - 수정 불가  >> 관리자가 답변을 달아 완료가 된 상태인데 수정하려는 경우
         if(ask.getStatus() == StatusCategory.COMPLETION){
+            throw new CustomException(ErrorCode.NOT_MODIFICATION);
+        }
+
+        //403 - 권한 없음  >> 자기가 쓴 글이 아닌경우
+        if (!ask.getUser().getEmail().equals(email)) {
             throw new CustomException(ErrorCode.NO_AUTH);
         }
+
+
         try{
+
             Optional<Ask> modifyAsk = askRepository.findByAskId(request.getAskId());
             modifyAsk.get().setTitle(request.getTitle());
             modifyAsk.get().setContents(request.getContents());
@@ -138,11 +163,14 @@ public class AskService {
     }
 
 
-    private Ask findByAskId(Long askId) {
+    public  Ask findByAskId(Long askId) {
         return askRepository.findByAskId(askId)
                 .orElseThrow(()-> new CustomException(ErrorCode.NOT_EXIST_ID));
     }
 
-
+    public User findByEmail(String email) {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_EXIST_MEMBER));
+    }
 
 }
